final commit

+++ Assets/Scripts/Bullet/BulletSettings.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 public class BulletSettings : ScriptableObject
 {
     [SerializeField] private float speed;
-    [SerializeField] private float damage;
+    [SerializeField] private int damage;
     [SerializeField] private bool canPierce; // goes through enemies
 
     public float Speed { get { return speed; } }
-    public float Damage { get { return damage; } }
+    public int Damage { get { return damage; } }
     public bool CanPierce { get { return canPierce; } }
 }
